+++ docker/dokuwiki/local.php
@@ -1,6 +1,9 @@
 <?php
 
-$conf['title'] = 'DokuWiki Chordsheets Test';
+$conf['title'] = 'DokuWiki Chordsheets Demo';
+$conf['tagline'] = 'Interactive chord sheets for DokuWiki';
 $conf['lang'] = 'en';
 $conf['license'] = 'cc-by-sa';
 $conf['useacl'] = 0;
+$conf['breadcrumbs'] = 0;
+$conf['youarehere'] = 0;